2021リニューアル資材リリース

+++ back/src/main/java/jp/go/ndl/lab/back/domain/Illustration.java
@@ -39,6 +39,6 @@ public class Illustration implements EsData {
     }
     
     
-    public double[] feature;
+    public float[] feature;
 
 }
